feat: adding more unittests and clean up parser code
