+++ dojo/unittests/scans/gitlab_api_fuzzing/gitlab_api_fuzzing_invalid.json
@@ -1,0 +1,19 @@
+{
+	"version": "3.0",
+	"remediations": [],
+	"scan": {
+		"scanner": {
+			"id": "libfuzzer",
+			"name": "libfuzzer",
+			"url": "https://docs.gitlab.com/ee/user/application_security/coverage_fuzzing",
+			"vendor": {
+				"name": "GitLab"
+			},
+			"version": "f248ff98"
+		},
+		"type": "coverage_fuzzing",
+		"start_time": "2021-02-08T12:19:48",
+		"end_time": "2021-02-08T12:19:54",
+		"status": "success"
+	}
+  }
